% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexconn_normalize_image.R
\name{flexconn_normalize_image}
\alias{flexconn_normalize_image}
\title{Normalize Image}
\usage{
flexconn_normalize_image(image, contrast = c("T1", "T2", "FLAIR", "PD"))
}
\arguments{
\item{image}{Object of class \code{nifti} or filename}

\item{contrast}{Type of imaging modality.}
}
\value{
An object of class \code{nifti}
}
\description{
Normalize Image
}
