% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_paths.R
\name{get_model_paths}
\alias{get_model_paths}
\title{Getting Model Paths for trained models}
\usage{
get_model_paths(num_atlases = c("21", "61"), outcomes = c("mask1",
  "mask2"))
}
\arguments{
\item{num_atlases}{Specifies which model to use.
Determined by the number of atlases in the FLEXCONN model.}

\item{outcomes}{The outcome used to train the model, from rater 1
or rater 2}
}
\value{
A matrix of filenames
}
\description{
Getting Model Paths for trained models
}
\examples{
get_model_paths()
get_model_paths(num_atlases = "21")
get_model_paths(outcomes = "mask1")
get_model_paths(num_atlases = "21", outcomes = "mask1")
}
