% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_flexconn_reqs.R
\name{install_flexconn_reqs}
\alias{install_flexconn_reqs}
\title{Install FLEXCONN requirements}
\usage{
install_flexconn_reqs(gpu = FALSE, ...)
}
\arguments{
\item{gpu}{Install GPU tensorflow}

\item{...}{Additional arguments to pass to \code{\link{install_tensorflow}}}
}
\description{
Install FLEXCONN requirements
}
