% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_flexconn.R
\name{train_flexconn}
\alias{train_flexconn}
\title{Training for Fast Lesion Extraction using Convolutional Neural Networks (FLEXCONN)}
\usage{
train_flexconn(atlas_dir, patch_size = c(35, 35), outdir = NULL,
  gpu = "gpu", verbose = TRUE)
}
\arguments{
\item{atlas_dir}{Atlas directory containing atlasXX_T1.nii,
atlasXX_FL.nii.gz, atlasXX_mask.nii.gz, where \code{XX=1,2,3, etc}
All atlas images must be in axial RAI orientation, or whatever orientation
FLAIR has the highest in-plane resolution.  Atlas T1 and FLAIR images must
be coregistered and have same dimensions.Z}

\item{patch_size}{Patch size, e.g. 35x35 or 31x31 (2D).
Patch sizes are separated by x.
Note that 2D patches are employed because usually FLAIR images are acquired 2D.}

\item{outdir}{Output directory where the trained models are written.}

\item{gpu}{Choice for GPU. Use the integer ID for the GPU.
Use "cpu" to use CPU.}

\item{verbose}{Print diagnostic messages}

\item{num_atlases}{Specifies which model to use.
Determined by the number of atlases in the FLEXCONN model.}
}
\value{
A vector of filenames
}
\description{
Training for Fast Lesion Extraction using Convolutional Neural Networks (FLEXCONN)
}
\examples{
# predict_flexconn(python_cmd = "python3)
library(reticulate)
use_python("/Library/Frameworks/Python.framework/Versions/3.5/bin/python3")
flair = system.file("extdata", "FLAIR.nii.gz", package = "flexconnr")
t1 = system.file("extdata", "T1.nii.gz", package = "flexconnr")
pp = predict_flexconn(t1 = t1, flair = flair)

}
