% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexconn_normalize_image.R
\name{flexconn_normalize_image}
\alias{flexconn_normalize_image}
\title{Normalize Image}
\usage{
flexconn_normalize_image(image, contrast = c("T1", "T2", "FLAIR", "PD"))
}
\arguments{
\item{image}{Object of class \code{nifti} or filename}

\item{contrast}{Type of imaging modality.}
}
\value{
An list of object of class \code{nifti} and the peak used to norm.
}
\description{
Normalize Image
}
\examples{
\dontrun{
library(reticulate)
use_python("/Library/Frameworks/Python.framework/Versions/3.5/bin/python3")

flair = system.file("extdata", "FLAIR.nii.gz", package = "flexconnr")
norm_flair = flexconn_normalize_image(flair, "FLAIR")
}
}
